+++ GameShopLibrary/PayOneGameItem.cs
@@ -4,7 +4,7 @@ namespace GameShopLibrary
 {
     class PayOneGameItem : Item
     {
-        public PayOneGameItem(int cost, Guid id, Game game) : base (cost, id, game)
+        public PayOneGameItem(double cost, Guid id, Game game) : base (cost, id, game)
         {
 
         }
